refactor(customer): extract recent route widgets
