+++ apps/customer/lib/widgets/recent_route_action.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class RecentRouteAction extends StatelessWidget {
+  const RecentRouteAction({
+    super.key,
+    required this.onPressed,
+  });
+
+  final VoidCallback onPressed;
+
+  @override
+  Widget build(BuildContext context) {
+    return OutlinedButton(
+      onPressed: onPressed,
+      style: OutlinedButton.styleFrom(
+        minimumSize: const Size(0, 42),
+        padding: const EdgeInsets.symmetric(horizontal: 14),
+      ),
+      child: const Text('Rebook'),
+    );
+  }
+}
